Refactor home and test screens
